requirements.txt
streamlit>=1.31.0
langchain>=0.1.16
langchain-community>=0.0.29
langchain-openai>=0.0.8
openai>=1.12.0
python-dotenv>=1.0.0